vsp-renderer: set default BRU input formats
BRU shall have same formats on input and output side. However, when
the system starts up, the default is set to YUV. Furthermore,
input formats shall be set before the output format, i.e. if the input
formats are not set to RGB, the output format won't be set to RGB
neither.

As we use RGB composition only, we set the input format to be RGB
at the startup time.