v4l2-renderer: fix transformation precisions.
When the sub-pixel pixel coordinates are given, the value were truncated
to the nearest integer value. This made width and height smaller than
they're supposed to be.

The default weston-scaler should now show white border on the right.

We also fixes the debug messages for transformation matrix to show
decimals as well.