v4l2-renderer: gl-fallback: defer attach buffer in gl-renderer
Attaching buffer in gl-renderer may consistenly become an overhead if we
just compose with v4l2-renderer. By deferring the buffer attachment in
gl-renderer until we actually fallback to the gl-renderer, we reduce the
amount of overhead.

On the otherhand, latencies increase when we switch back and forth
between v4l2-renderer and gl-renderer.