v4l2-renderer: gl-fallback: check the current renderer_state in the weston_surface
When falling back to gl-renderer, the renderer_state in the surface
needs to be changed to that of gl-renderer's. And then set back to the
v4l2-renderer's renderer_state after gl-renderer is done with it.
We were doing this by iterating through views passed to the renderer.

When a surface is referred from the multiple views, this iteration
caused the problem. We have to make sure that we are setting the
renderer_state of the surface only once.